perf(lgn): fix T4 OOM by switching LogicLayer to polynomial form
poc_a_lgn OOM'd on T4 because each LogicLayer materialized a
(B, T, n, 16) = 1 GiB fp32 stack tensor and 16 of them stayed live for
backward. Every binary gate is multilinear of degree ≤ 2 in (a, b), so
Σ_g p_g · gate_g(a, b) collapses to α + β·a + γ·b + δ·a·b where
(α, β, γ, δ) = p @ GATE_COEFFS. Output is numerically identical (asserted
in a new test) but never builds the 16-stack — ~16× less LGN activation
memory.

Also cast ThermometerEncode output to x.dtype: theta and s are fp32
Parameters that promote bf16 inputs back to fp32, which would silently
defeat autocast for the entire LGN body if the residual stream is ever
moved to bf16.

+++ nanolgn/lgn_mlp.py
@@ -47,8 +47,11 @@ class ThermometerEncode(nn.Module):
         x_e = x.unsqueeze(-1)                          # (..., d, 1)
         # theta: (d, K) → broadcasts cleanly.
         # s: (K,) → broadcasts cleanly.
+        # Params are fp32 even under autocast(bf16), which would otherwise
+        # promote the entire LGN body to fp32 and quadruple activation memory.
+        # Cast back to x's dtype to keep the body in the autocast dtype.
         b = torch.sigmoid(self.s * (x_e - self.theta))  # (..., d, K)
-        return b.flatten(-2)                           # (..., d*K)
+        return b.flatten(-2).to(x.dtype)               # (..., d*K)
 
 
 class GroupSumDecode(nn.Module):
